pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sync"]

[project]
name = "wpt-sync"
requires-python = ">=3.12"
description = "Gecko sync bot for web-platform-tests"
authors = [
  {name = "Mozilla Corporation", email = "mozilla-tools@lists.mozilla.org"}
]
version = "0.2.0"
classifiers = [
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "bugsy==0.12.0",
  "celery==5.5.3",
  "filelock==3.18.0",
  "gitpython==3.1.43",
  "kombu==5.5.4",
  "lando-cli==0.0.9",
  "mercurial==7.0.2",
  "mozautomation",
  "newrelic==10.13.0",
  "phabricator==0.9.1",
  "pygit2==1.18.0",
  "pygithub==2.6.1",
  "pyyaml==6.0.2",
  "requests==2.32.3",
  "setuptools==80.9.0",
  "slugid==2.0.0",
  "taskcluster==84.2.0",
]


[project.optional-dependencies]
test = [
  "celery_stubs==0.1.3",
  "mypy==1.16.1",
  "pytest==8.3.3",
  "pytest-cov==5.0.0",
  "pytest-env==1.1.5",
  "requests-mock==1.12.1",
  "ruff==0.12.1",
  "types-filelock==3.2.7",
  "types-requests==2.32.0.20250306",
  "types-PyYAML==6.0.12.20250516"
]
deploy = [
  "ansible==8.5.0"
]

[project.scripts]
wptsync = 'sync.command:main'

[tool.uv.sources]
mozautomation = { git = "https://github.com/mozilla/version-control-tools.git", subdirectory = "pylib/mozautomation", rev = "c40d7a3be799b3c23ebe6289def25bef0f51893a" }

[tool.ruff]
line-length = 100

[tool.mypy]
files = "sync/**/*.py"
check_untyped_defs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "bugsy.*,newrelic.*,lib2to3.*,pygit2.*,kombu.*,mozautomation.*,phabricator.*,slugid.*,taskcluster.*"

[tool.pytest.ini_options]
env = [
    "WPTSYNC_ROOT=/app/testdata",
    "WPTSYNC_REPO_ROOT=/app/testdata"
]
